refactor: move server tools into server/tools module
